+++ iqm_station_control_client/src/iqm/station_control/client/qon.py
@@ -292,7 +292,7 @@ class QONMetric(QON):
         "irb",
         "ssro",
         "restless_ssro",
-        "readout_rb",
+        "rrc",
         "qndness",
         "msmt_qndness",
         "npopex_echo",
@@ -735,4 +735,12 @@ class ObservationFinder(dict):
             return _convert_to_float(error_0_to_1), _convert_to_float(error_1_to_0)
         except KeyError:
             logger.warning("Missing errors for %s.%s.%s.", gate_name, impl_name, locus_str)
+            return None
+
+    def get_qubit_frequency(self, qubit: str) -> float | None:
+        """Qubit drive frequency, or None if not found."""
+        try:
+            return self._get_path_value(["controllers", qubit, "drive", "frequency"])
+        except KeyError:
+            logger.warning(f"Missing drive frequency for {qubit}.")
             return None