+++ iqm_station_control_client/src/iqm/station_control/interface/models/__init__.py
@@ -27,6 +27,7 @@ from iqm.station_control.interface.models.jobs import (
     TimelineEntry,
 )
 from iqm.station_control.interface.models.observation import (
+    ObservationBase,
     ObservationData,
     ObservationDefinition,
     ObservationLite,