+++ iqm_client/CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog
 =========
 
+Version 29.14.0 (2025-08-20)
+============================
+
+Bug fixes
+---------
+
+- Add explicit cross-component requirements
+
+Version 29.13.0 (2025-08-19)
+============================
+
+Bug fixes
+---------
+
+- Fix ``iqm-client[cirq]`` installation for all supported Python versions.
+
 Version 29.12.0 (2025-08-08)
 ============================